Add new implementation of Threshold Printer
Again, still not wired up. Implementation is fully tested, but will not be used
to replace existing solution until HTML printer is also implemented, so we can
replace all existing output strategies (save the underused and underloved matrix
output strategy).

+++ src/main/java/edu/wpi/checksims/algorithm/output/SimilarityMatrix.java
@@ -103,8 +103,8 @@ public class SimilarityMatrix {
                 throw new RuntimeException("Could not find index of submission " + result.b.getName());
             }
 
-            results[indexFirst][indexSecond] = result.percentMatchedA();
-            results[indexSecond][indexFirst] = result.percentMatchedB();
+            results[indexFirst][indexSecond] = (float)result.percentMatchedA();
+            results[indexSecond][indexFirst] = (float)result.percentMatchedB();
         });
 
         logs.info("Done performing similarity detection");
